+++ react-app/src/components/NavBar/NavBar.css
@@ -1,0 +1,14 @@
+nav {
+    position: sticky;
+    top: 0;
+    width: 100%;
+    max-width: 100%;
+    background-color: var(--gray-primary);
+    border-bottom: 2px solid var(--white);
+}
+
+.nav-container {
+    height: 7.2rem;
+    padding: 1rem 2.5rem;
+
+}
