+++ react-app/src/components/auth/AuthForm.css
@@ -1,6 +1,6 @@
 .auth-form-page {
     width: 100%;
-    height: 100%;
+    height: 75%;
     display: flex;
     align-items: center;
     justify-content: center;
@@ -53,7 +53,7 @@
 }
 
 .auth-form button:hover {
-    background-color: var(--black);
+    background-color: var(--gray-primary);
 }
 
 .auth-form button:disabled {
@@ -97,4 +97,12 @@
 
 .repeat-input-invalid {
     border-bottom: .3rem solid var(--orange-secondary)!important;
+}
+
+.signup-form-container {
+    margin-top: 19rem;
+}
+
+.login-form-container {
+    margin-top: 10rem;
 }
